requirements.txt
streamlit>=1.45.0,<2.0.0
openai>=1.57.1,<2.0.0
pandas>=2.2.3,<3.0.0
groq>=0.13.0,<1.0.0
pydantic>=2.10.11,<3.0.0
pyyaml>=6.0.1,<7.0.0
python-dotenv>=1.0.1,<2.0.0
pypdf>=4.3.1,<5.0.0
couchbase>=4.1.0
google-genai
google-adk